ISSUE #478 a lot of rewrite to make this work - seems to be working for 2x3 now.
