+++ bouncer/src/repo/manipulator/modelconvertor/import/repo_model_import_ifc.cpp
@@ -58,6 +58,7 @@ bool IFCModelImport::importModel(std::string filePath, uint8_t &err)
 
 	repoInfo << "IMPORT [" << fileName << "]";
 	repoInfo << "=== IMPORTING MODEL WITH IFC OPEN SHELL ===";
+
 	bool success = false;
 	std::string errMsg;
 	ifcHelper::IFCUtilsGeometry geoUtil(filePath, settings);
